requirements/prod.txt
-r common.txt
uWSGI==2.0.12
Flask-SSLify==0.1.5
